v6.28
Removed reset from sub vi - AI USB4704 (3).vi - Presumably this won't work if the Task is using that device!

There is an existing  USB Reset routine which appeared to never be called. The Analog Error variable has been connected to the fault O/P of AI vi. Hopefully this will trigger a reset